% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{datatype}
\alias{datatype}
\alias{gdal_datatypes}
\title{Return the type name of the GDAL data type.}
\usage{
gdal_datatypes()

datatype(x)
}
\arguments{
\item{x}{integer as returned by GDAL, or osgeo.gdal.Open().GetDatatype()}
}
\value{
character string of the type name (the name of the constant in GDAL, e.g. GDT_Byte)
}
\description{
Return the type name of the GDAL data type.
}
\examples{
datatype(1)
names(gdal_datatypes())
}
