% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsn-package.R
\docType{package}
\name{dsn-package}
\alias{dsn}
\alias{dsn-package}
\title{dsn: Data Source Name and Description Helpers for Use with 'GDAL'}
\description{
Simple helpers for 'GDAL' data source names ('DSN'), prefix and suffix and other handling. 'GDAL' is the Geospatial Data Abstraction Library \url{https://gdal.org/}, not used by this package directly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/dsn}
  \item \url{https://hypertidy.github.io/dsn/}
  \item Report bugs at \url{https://github.com/hypertidy/dsn/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Michael Sumner \email{mdsumner@gmail.com}
}

Other contributors:
\itemize{
  \item Matt Dowle (original idea for C_addr, from data.table::address) [contributor]
}

}
\keyword{internal}
