% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{mursst}
\alias{mursst}
\title{MURSST (GHRSST) sst Zarr source}
\usage{
mursst(band = 0)
}
\arguments{
\item{band}{a band number (defaults to 0, which is 2002-06-01)}
}
\value{
a string, a dsn for Zarr MURSST
}
\description{
Don't trust this, I don't know if it's really daily like the NetCDF files are. GDAL should have the time metadata in the 2D interface but does not.
It seems like this source doesn't go past "2020-01-21" or band 6443, don't know why that is.
}
\examples{
mursst()
mursst_time("2019-10-08")
}
