% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{sds}
\alias{sds}
\title{Wrapping handlers for GDAL data source names}
\usage{
sds(x, varname, driver, quote = TRUE)
}
\arguments{
\item{x}{character vector, of data source names (file paths, urls, database connection strings, or GDAL dsn)}

\item{varname}{named of variable in DSN}

\item{driver}{driver to use, e.g. "NETCDF", "HDF5"}

\item{quote}{wrap the core dsn in escaped double quotes, or not}
}
\value{
character string of the form "DRIVER:\%s:varname"
}
\description{
Subdataset and VRT connection strings.
}
\examples{
f <- "myfile.nc"
sds(f, "variable", "NETCDF", quote = FALSE)
}
