% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsn-package.R
\docType{package}
\name{dsn-package}
\alias{dsn}
\alias{dsn-package}
\title{dsn: Data Source Name and Description Helpers for Use With 'GDAL'}
\description{
Simple helpers for 'GDAL' data source names ('DSN'), prefix and suffix and other handling. 'GDAL' is the Geospatial Data Abstraction Library, not used by this package directly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/dsn}
  \item Report bugs at \url{https://github.com/hypertidy/dsn/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Matt Dowle (wrote the source code used here in C_addr, originally in data.table) [contributor]
}

}
\keyword{internal}
