% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcp_extent}
\alias{gcp_extent}
\alias{gcp_extent_arg}
\title{Create a set of GCPs (ground control points) from dimension, extent}
\usage{
gcp_extent(dimension, extent = NULL)

gcp_extent_arg(gcp)
}
\arguments{
\item{dimension}{size of grid 'ncol,nrow'}

\item{extent}{extent 'xmin,xmax,ymin,ymax'}

\item{gcp}{ground control point ('col,row,x,y')}
}
\value{
\code{gcp_extent} returns the col,row,x,y values, \code{gcp_extent_arg} returns
formatted as a GDAL 'vrt://' connection string
}
\description{
Create a set of GCPs (ground control points) from dimension, extent
}
\examples{
gcp_extent(c(10, 20))
dsn <- sprintf("vrt://\%s?\%s", mem(volcano), gcp_extent_arg(gcp_extent(dim(volcano))))
gcp <- gcp_extent(dim(volcano), c(-180, 180, -90, 90))
gcp_extent_arg(gcp)
}
