% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{prefix}
\alias{prefix}
\alias{vsicurl}
\alias{driver}
\alias{netcdf}
\alias{unprefix}
\alias{unvsicurl}
\title{Prefix handlers for GDAL data source names}
\usage{
vsicurl(x)

driver(x, driver = "")

netcdf(x)

unprefix(x)

unvsicurl(x)
}
\arguments{
\item{x}{character vector, of data source names (file paths, urls, database connection strings, or GDAL dsn)}

\item{driver}{character vector of appropriate GDAL driver name}
}
\value{
character vector
}
\description{
Add required prefixes, or remove them.
}
\examples{
vsicurl("https://netcdf-r-us.org/f.nc")

driver("somefile.h5", "HDF5")

unvsicurl("/vsicurl/https://netcdf-r-us.org/f.nc")

unprefix("NETCDF:/u/user/somefile.nc")
}
