% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{vrtcon}
\alias{vrtcon}
\title{VRT connection}
\usage{
vrtcon(x, ...)
}
\value{
character string in the form "vrt://\%s?arg1&arg2"
}
\description{
Create a vrt connection from an input string and named arguments.
}
\details{
As of writing (GDAL 3.7.0DEV 2022-12-12) the only available named arguments
are 'a_srs', 'bands', 'a_ullr' but that doesn't stop this function.
}
\examples{

vrtcon("myfile.nc", a_ullr = "0,90,360,-90", bands="1,2,1")
}
